+++ packages/core-components/src/types/mdx-components.d.ts
@@ -14,6 +14,6 @@
 import type { MDXComponents } from "mdx/types";
 
 declare const mdxComponents: MDXComponents;
-export default mdxComponents;
+export { mdxComponents };
 
 export declare function useMDXComponents(components: MDXComponents): MDXComponents;